Add bad gravity.

+++ src/gfx.rs
@@ -6,7 +6,7 @@ const MAX_BRIGHTNESS: u8 = 255;
 const DISPERSION_MATRIX_SIZE: u8 = 9;
 const DISPERSED: [u8; DISPERSION_MATRIX_SIZE as usize] = [1, 7, 4, 5, 8, 3, 6, 2, 9];
 
-pub fn draw_pixel(x: u32, y: u32, brightness: u8) {
+pub fn draw_pixel(x: u8, y: u8, brightness: u8) {
     if x > 160 || y > 160 {
         return;
     }
@@ -15,7 +15,7 @@ pub fn draw_pixel(x: u32, y: u32, brightness: u8) {
 
     let color_base = brightness / step;
     // Get the value from the matrix.
-    let idx = ((x - y * 3) % u32::from(DISPERSION_MATRIX_SIZE)) as usize;
+    let idx = ((x - y * 3) % DISPERSION_MATRIX_SIZE) as usize;
     let threshhold = unsafe { DISPERSED.get_unchecked(idx) * (step / DISPERSION_MATRIX_SIZE) };
     let diff = brightness % step;
     let add_color = if diff >= threshhold { 1 } else { 0 };
@@ -28,7 +28,7 @@ pub fn draw_pixel(x: u32, y: u32, brightness: u8) {
     screen_pixel(x, y);
 }
 
-fn screen_pixel(x: u32, y: u32) {
+fn screen_pixel(x: u8, y: u8) {
     // The byte index into the framebuffer that contains (x, y)
     let idx = (y as usize * 160 + x as usize) >> 2;
 
@@ -36,6 +36,7 @@ fn screen_pixel(x: u32, y: u32) {
     let shift = (x as u8 & 0b11) << 1;
     let mask = 0b11 << shift;
 
+    #[allow(clippy::indexing_slicing)]
     unsafe {
         let palette_color: u8 = (*wasm4::DRAW_COLORS & 0xf) as u8;
         if palette_color == 0 {
